requirements.txt
tensorflow-cpu==2.17.0
opencv-python
ultralytics
streamlit==1.35.0
streamlit-login-auth-ui==0.2.0
trycourier==4.2.0
argon2-cffi==23.1.0
